+++ Quiz-Website/src/main/java/com/freeuni/proj_100/quizwebsite/controller/QuizController.java
@@ -33,7 +33,7 @@ public class QuizController {
      * Accessible at: GET /quiz/{id}
      */
     @GetMapping("/{id}")
-    public String quizPage(@PathVariable Long id, Model model) {
+    public String quizPage(@PathVariable Integer id, Model model) {
         Optional<Quiz> quiz = quizService.getQuizById(id);
         if (quiz.isEmpty()) {
             return "redirect:/";
@@ -47,7 +47,7 @@ public class QuizController {
      * Accessible at: GET /quiz/{id}/take
      */
     @GetMapping("/{id}/take")
-    public String takeQuiz(@PathVariable Long id, Model model) {
+    public String takeQuiz(@PathVariable Integer id, Model model) {
         Optional<Quiz> quiz = quizService.getQuizById(id);
         if (quiz.isEmpty()) {
             return "redirect:/";
@@ -61,7 +61,7 @@ public class QuizController {
      * Accessible at: POST /quiz/{id}/submit
      */
     @PostMapping("/{id}/submit")
-    public String submitQuiz(@PathVariable Long id,
+    public String submitQuiz(@PathVariable Integer id,
                              @RequestParam java.util.Map<String, String> answers,
                              Model model) {
         Optional<Quiz> quiz = quizService.getQuizById(id);
